Remove Optional versions of `<|` operators
The addition of overloads for `<|` and friends that return optional
values sounded like a great idea but I think they have ended up being
awkward in practice.

The problem that they solve is, frankly, minimal. It helps when decoding
from a root key, but that's about it.

On the other hand, introducing an entirely new set of operators adds
overhead and introduces possible ambiguity for the compiler as well as
introduces cognitive overhead for the user. I'm not convinced that some
of the issues we've seen around expression complexity and poor error
messaging aren't due to the introduction of these (rarely used)
operators.

The _real_ solution would be for Swift to support Monads and then we
could make these completely generic, but I'm not holding my breath
there.